build: change docker compose exec to run
Agreed to change from `docker compose exec` to `docker compose run`
after discussion with @spacesnottabs . This allows the sleep entrypoint
to be removed. This wasn't convenient previously, because of the need to
be able to change Bundler settings to point to local gems. Since that
configuration is now automatic on boot and controlled via environment
variables, this change is now possible without inconvenience.