example latest env vars
The latest env vars running in Heroku and EC2 are a bit out of sync with
the exampled ones. Update them, to decrease distance between dev and
other environments, and to have a record of the latest config.

This is pretty important for this repo, since I'd never be able to guess
the settings otherwise.