[#19] extend Companies to add industry_code_uids
Note that it was necessary to filter the array in Athena, since
otherwise there were no empty arrays, only arrays containing the empty
string as a single element. I'm not sure if that's normal behaviour or
not, but it makes querying both surprising and more difficult, so I
filtered it.